requirements.txt
playwright==1.49.1
openai==1.82.0
gspread==6.1.4
google-auth==2.38.0
python-telegram-bot==21.10
python-dotenv==1.0.1
